+++ core/index.commonjs.js
@@ -8,6 +8,8 @@ exports.parsePhoneNumberFromString = require('../build/parsePhoneNumberFromStrin
 
 exports.findNumbers = require('../build/findNumbers').default
 exports.searchNumbers = require('../build/searchNumbers').default
+exports.findPhoneNumbersInText = require('../build/findPhoneNumbersInText').default
+exports.searchPhoneNumbersInText = require('../build/searchPhoneNumbersInText').default
 exports.PhoneNumberMatcher = require('../build/PhoneNumberMatcher').default
 
 exports.AsYouType = require('../build/AsYouType').default