+++ test/exports.test.js
@@ -23,6 +23,8 @@ import
 
 	findNumbers,
 	searchNumbers,
+	findPhoneNumbersInText,
+	searchPhoneNumbersInText,
 	PhoneNumberMatcher,
 
 	AsYouType,
@@ -87,6 +89,8 @@ describe(`exports`, () => {
 
 		findNumbers('+12133734253', 'US').should.deep.equal([{ country: 'US', phone: '2133734253', startsAt: 0, endsAt: 12 }])
 		searchNumbers('+12133734253', 'US')[Symbol.iterator]().next.should.be.a('function')
+		findPhoneNumbersInText('+12133734253')[0].number.number.should.equal('+12133734253')
+		searchPhoneNumbersInText('+12133734253')[Symbol.iterator]().next.should.be.a('function')
 		new PhoneNumberMatcher('+12133734253', undefined).find.should.be.a('function')
 
 		new AsYouType('US').input('+12133734253').should.equal('+1 213 373 4253')
@@ -151,6 +155,8 @@ describe(`exports`, () => {
 
 		Library.findNumbers('+12133734253', 'US').should.deep.equal([{ country: 'US', phone: '2133734253', startsAt: 0, endsAt: 12 }])
 		Library.searchNumbers('+12133734253', 'US')[Symbol.iterator]().next.should.be.a('function')
+		Library.findPhoneNumbersInText('+12133734253')[0].number.number.should.equal('+12133734253')
+		Library.searchPhoneNumbersInText('+12133734253')[Symbol.iterator]().next.should.be.a('function')
 		new Library.PhoneNumberMatcher('+12133734253', undefined).find.should.be.a('function')
 
 		new Library.AsYouType('US').input('+12133734253').should.equal('+1 213 373 4253')